+++ esperanto-tools-libs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Conan) Depend on sw-sysemu/0.2.1
 - [SW-14434] (Conan) CI now generates Debug & Release pkgs
 - *** Breaking Change *** changed ArchRevision from DeviceProperties to be consistent with DeviceLayer
+- Adapted to deviceLayer breaking changes
 ### Deprecated
 ### Removed
 ### Fixed